LayoutTests/http/wpt/beacon/connect-src-beacon-redirect-allowed.sub-expected.txt


PASS Redirect is allowed by CSP
